add Test class, IBookStoreDBContext and tests

+++ WebApi/Application/AuthorOperations/Commands/DeleteAuthor/DeleteAuthorCommand.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Linq;
 using AutoMapper;
 using WebApi.DBOperations;
 using WebApi.Entities;
@@ -9,9 +10,9 @@ namespace WebApi.Application.AuthorOperations.Commands.DeleteAuthor
     {
         public int ID { get; set; }
 
-        private readonly BookStoreDbContext _dbContext;
+        private readonly IBookStoreDbContext _dbContext;
 
-        public DeleteAuthorCommand(BookStoreDbContext dbContext)
+        public DeleteAuthorCommand(IBookStoreDbContext dbContext)
         {
             _dbContext = dbContext;
         }
@@ -19,9 +20,15 @@ namespace WebApi.Application.AuthorOperations.Commands.DeleteAuthor
         public void Handle()
         {
             Author author = _dbContext.Authors.Find(ID);
+            
             if(author is null)
             {
                 throw new InvalidOperationException("Silinecek Yazar bulunamadı");
+            }
+            
+            if(_dbContext.Books.Any(a=>a.AuthorID == ID))
+            {
+                throw new InvalidOperationException("Yazarı Silmeden Önce Yazarın Kitaplarını Siliniz!");
             }
             _dbContext.Authors.Remove(author);
             _dbContext.SaveChanges();
